papers/quantum_transfer_learning/requirements.txt
# Quantum Transfer Learning - Dependencies
# Install with: pip install -r requirements.txt

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Quantum computing - MerLin (photonic)
merlinquantum==0.2.3
perceval-quandela>=0.11.0

# Quantum computing - PennyLane (qubit-based comparison)
pennylane>=0.33.0

# Data processing
scikit-learn>=1.3.0
pandas>=2.0.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0

# Testing
pytest>=7.0.0
